python-ecosys/aiohttp: Add timeout attribute to ClientSession.
Implement a timeout while awaiting ClientSession reader.
in __init__.py add a _timeout attribute with a default of `None`
amend method _request when awaiting reader to use an asyncio.wait_for.
The aiohttp.ClientSession can then be created with an optional timeout,
the caller can optionally catch the asyncio.TimeoutError